Restructured .js files in src, folder structure should now be better Removed a few unused files and imports
