+++ src/components/notes/NoteView.js
