+++ src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './css/Index.css';
 import '../node_modules/font-awesome/css/font-awesome.min.css';
-import App from './scripts/App';
-import registerServiceWorker from './scripts/registerServiceWorker';
+import App from './components/App';
+import registerServiceWorker from './components/registerServiceWorker';
 
 ReactDOM.render(<App />, document.getElementById('root'));
 registerServiceWorker();
