--- src/scripts/List.js
@@ -1,10 +1,0 @@
-import React from 'react';
-const List = props => (
-  <ul className="list">
-  {
-    props.items.map((item, index) => <li key={index}>{item}</li>)
-  }
-  </ul>
-);
-
-export default List;
